address PR comments: use task_system_handle for enumerable_thread_specific
