+++ arbor/thread_private_spike_store.cpp
@@ -3,6 +3,7 @@
 #include <arbor/common_types.hpp>
 #include <arbor/spike.hpp>
 
+#include "threading/enumerable_thread_specific.hpp"
 #include "threading/threading.hpp"
 #include "thread_private_spike_store.hpp"
 
@@ -11,14 +12,16 @@ namespace arb {
 struct local_spike_store_type {
     threading::enumerable_thread_specific<std::vector<spike>> buffers_;
 
-    local_spike_store_type(const task_system_handle& ts): buffers_(*ts) {};
+    local_spike_store_type(const task_system_handle& ts): buffers_(ts) {};
 };
 
-thread_private_spike_store::thread_private_spike_store(thread_private_spike_store&& t): impl_(std::move(t.impl_)) {};
+thread_private_spike_store::thread_private_spike_store(thread_private_spike_store&& t):
+    impl_(std::move(t.impl_))
+{}
 
 thread_private_spike_store::thread_private_spike_store(const task_system_handle& ts):
-    impl_(new local_spike_store_type(ts)) {
-}
+    impl_(new local_spike_store_type(ts))
+{}
 
 thread_private_spike_store::~thread_private_spike_store() {}
