+++ arbor/threading/threading.hpp
@@ -95,56 +95,6 @@ public:
     std::unordered_map<std::thread::id, std::size_t> get_thread_ids() const;
 };
 
-///////////////////////////////////////////////////////////////////////
-// types
-///////////////////////////////////////////////////////////////////////
-
-template <typename T>
-class enumerable_thread_specific {
-    std::unordered_map<std::thread::id, std::size_t> thread_ids_;
-
-    using storage_class = std::vector<T>;
-    storage_class data;
-
-public:
-    using iterator = typename storage_class::iterator;
-    using const_iterator = typename storage_class::const_iterator;
-
-    enumerable_thread_specific(const task_system& ts):
-        thread_ids_{ts.get_thread_ids()},
-        data{std::vector<T>(ts.get_num_threads())}
-    {}
-
-    enumerable_thread_specific(const T& init, const task_system& ts):
-        thread_ids_{ts.get_thread_ids()},
-        data{std::vector<T>(ts.get_num_threads(), init)}
-    {}
-
-    T& local() {
-        return data[thread_ids_.at(std::this_thread::get_id())];
-    }
-    const T& local() const {
-        return data[thread_ids_.at(std::this_thread::get_id())];
-    }
-
-    auto size() const { return data.size(); }
-
-    iterator begin() { return data.begin(); }
-    iterator end()   { return data.end(); }
-
-    const_iterator begin() const { return data.begin(); }
-    const_iterator end()   const { return data.end(); }
-
-    const_iterator cbegin() const { return data.cbegin(); }
-    const_iterator cend()   const { return data.cend(); }
-};
-
-inline std::string description() {
-    return "CThread Pool";
-}
-
-constexpr bool multithreaded() { return true; }
-
 class task_group {
 private:
     std::atomic<std::size_t> in_flight_{0};
